requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
mysqlclient==2.2.0
Pillow==10.1.0
reportlab==4.0.8
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3